AO3-2839 Move devise into Users namespace
