+++ app/views/users/registrations/_legal.html.erb
