+++ app/views/users/registrations/delete_preview.html.erb
