+++ app/views/users/registrations/destroy.html.erb
