+++ app/views/users/registrations/new.html.erb
