+++ app/views/users/sessions/_greeting.html.erb
