+++ app/views/users/sessions/new.html.erb
@@ -36,7 +36,7 @@
 <% end %>
 
 <div id="loginform">
-  <%= render 'user/sessions/form' %>
+  <%= render 'users/sessions/form' %>
 </div>
 
 <p>
